+++ wavespectra/specdataset.py
@@ -6,7 +6,7 @@ import sys
 import xarray as xr
 
 from wavespectra.core.attributes import attrs
-from wavespectra.select import sel_idw, sel_nearest, sel_bbox
+from wavespectra.core.select import sel_idw, sel_nearest, sel_bbox
 from wavespectra.specarray import SpecArray
 
 here = os.path.dirname(os.path.abspath(__file__))
